pyproject.toml
[tool.poetry]
name = "coursework-5"
version = "0.1.0"
description = ""
authors = ["u_alex90 <u_alex90@mail.ru>"]
readme = "README.md"
packages = [{include = "coursework_5"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
oyaml = "^1.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.7"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.1"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
